+++ bin/src/main/resources/application.properties
@@ -1,0 +1,6 @@
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5InnoDBDialect
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.url=jdbc:mysql://localhost:3306/quiz
+spring.datasource.username=quartetoFantastico
+spring.datasource.password=123
