+++ src/main/java/br/com/game/service/UserService.java
@@ -24,4 +24,5 @@ public class UserService {
 		return userRepository.save(user);
 	}
 
+	
 }
